+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Removed
+- Micro Count widget.
+
+### Added 
+- Generic trustpilot widget block.
 
 ## [1.0.2] - 2020-05-19
 ### Fixed